Added: Method to Find Maximum of Threee Float Values

+++ src/main/java/MaximumCheck/MaximumCheck.java
@@ -4,8 +4,9 @@ import java.awt.print.Printable;
 
 public class MaximumCheck {
 
-	public Integer maximumValue(Integer x, Integer y, Integer z) {
-		Integer max = x;
+	//UC1: Method to check Maximum value from 3 Floating Values
+	public Float maximumValue(Float x, Float y, Float z) {
+		Float max = x;
 		if (y.compareTo(max) > 0) {
 			max = y;
 		} if (z.compareTo(max) > 0) {
